# If not avx2 nor macos, make the reduce1_vt call reduce1