Without AVX2 instructions, the `reduce1_vt` seems quite bad on `x86_64`, so we call `reduce1` instead.