@tlepoint 2023-10-13
CPU:
```bash
$ cat /proc/cpuinfo
processor       : 7
vendor_id       : GenuineIntel
cpu family      : 6
model           : 85
model name      : Intel(R) Xeon(R) Platinum 8124M CPU @ 3.00GHz
[...]
```

Command ran for benchmarks
```bash
$ cargo criterion ntt --package fhe-math
$ cargo criterion bfv/mul --package fhe
```

Before:
```
ntt/forward/1024/62     time:   [10.828 µs 10.848 µs 10.883 µs]                                
ntt/forward_vt/1024/62  time:   [15.813 µs 15.830 µs 15.851 µs]                                   
ntt/backward/1024/62    time:   [11.966 µs 12.024 µs 12.126 µs]                                 
ntt/backward_vt/1024/62 time:   [12.014 µs 12.062 µs 12.133 µs]
[...]
bfv/mul/n=1024/log(q)=27                                                                            
                        time:   [722.27 µs 723.07 µs 724.25 µs]
bfv/mul/n=2048/log(q)=54                                                                           
                        time:   [1.4890 ms 1.4941 ms 1.5008 ms]
bfv/mul/n=4096/log(q)=109                                                                           
                        time:   [6.2704 ms 6.2783 ms 6.2874 ms]
bfv/mul_then_relinearize/n=4096/log(q)=109                        
                        time:   [7.1263 ms 7.1348 ms 7.1440 ms]
bfv/mul_and_relin/n=4096/log(q)=109                                                                           
                        time:   [6.7412 ms 6.7455 ms 6.7505 ms]
```

After:
```
ntt/forward/1024/62     time:   [10.834 µs 10.846 µs 10.865 µs]                                
ntt/forward_vt/1024/62  time:   [10.561 µs 10.564 µs 10.567 µs]                                   
ntt/backward/1024/62    time:   [11.968 µs 12.001 µs 12.039 µs]                                 
ntt/backward_vt/1024/62 time:   [11.407 µs 11.411 µs 11.416 µs]  
[...]
bfv/mul/n=1024/log(q)=27                                                                            
                        time:   [660.59 µs 662.08 µs 665.67 µs]
bfv/mul/n=2048/log(q)=54                                                                           
                        time:   [1.3564 ms 1.3569 ms 1.3581 ms]
bfv/mul/n=4096/log(q)=109                                                                           
                        time:   [5.8097 ms 5.8183 ms 5.8304 ms]
bfv/mul_then_relinearize/n=4096/log(q)=109                        
                        time:   [6.6487 ms 6.6610 ms 6.6790 ms]
bfv/mul_and_relin/n=4096/log(q)=109                                                                           
                        time:   [6.2875 ms 6.2995 ms 6.3116 ms]
```

Note that when we enable `avx2` in `RUSTC`, it is better to use the variable time version.